requirements.txt
asyncpg==0.27.0
fastapi~=0.100.0
alembic==1.10.2
gunicorn==20.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
python-dotenv==1.0.0
SQLAlchemy==2.0.9
uvicorn==0.20.0
PyJWT~=1.7.1
pydantic==1.10.4
starlette~=0.27.0
httpx~=0.24.0
Authlib~=1.2.1
passlib~=1.7.4
requests~=2.31.0
redis~=4.5.1
auth~=0.5.3
itsdangerous~=2.1.2
Jinja2~=3.1.2
email-validator==1.3.1